What's new: + Арифметические операции над дробными числами + Кнопки "Равно" и "Сброс" + История операций (исправлены ошибки)

+++ app/src/main/java/com/skarapedulbuk/mysimplecalc/MainActivity.java
@@ -42,7 +42,7 @@ public class MainActivity extends AppCompatActivity implements MainContract.View
     }
 
     @Override
-    public void showResult(int result) {
+    public void showResult(double result) {
         currentTextView.setText(String.valueOf(result));
     }
