+++ app/src/main/java/com/skarapedulbuk/mysimplecalc/MainContract.java
@@ -2,7 +2,7 @@ package com.skarapedulbuk.mysimplecalc;
 
 public interface MainContract {
     interface View {
-        void showResult(int result);
+        void showResult(double result);
 
         void updateHistory(String history);
 
@@ -21,6 +21,6 @@ public interface MainContract {
     }
 
     interface Model {
-        int doOperation(int arg1, int arg2, Operations operator);
+        double doOperation(double arg1, double arg2, Operations operator);
     }
 }
